Added MFC Level start code and some Dart Tweaks
Added the beginning code for the level editor. Did two minor tweaks.
Moved using declarations in Level class from the header to the cpp and
changed it in the header to scope in, this is the best practice for
including namespace, scope in the header, using declaration in the cpp.
Lastly, I made it so that the window is centered on the screen.