Update: Found the actual problem
The reason why I had so much trouble with the code is that I downloaded
and installed the wrong version of FMOD. The newest version of FMOD did
not contain keywords like FMOD_CAPS or FMOD_CHANNEL_FREE. So it threw me
for a loop. Everything seems to be fixed, but there are warnings that
I'm currently trying to tackle.